src/tests/build/buildVerification.test.ts
import { BuildVerification } from './buildVerification';

describe('Build Verification', () => {
  let buildVerification: BuildVerification;

  const mockConfig = {
    environment: 'development' as const,
    requiredEnvVars: ['VITE_SUPABASE_URL', 'VITE_SUPABASE_ANON_KEY'],
    requiredDependencies: ['react', 'react-dom'],
    outputDir: './dist',
    staticAssets: ['assets/images', 'assets/css']
  };

  beforeEach(() => {
    buildVerification = new BuildVerification(mockConfig);
  });

  test('should verify environment variables', async () => {
    process.env.VITE_SUPABASE_URL = 'test-url';
    process.env.VITE_SUPABASE_ANON_KEY = 'test-key';
    
    const result = await buildVerification.verifyBuild();
    expect(result).toBeDefined();
  });

  test('should verify build output', async () => {
    const result = await buildVerification.verifyBuild();
    expect(result).toBeDefined();
  });
});